fix: rename project
